move fixture from test file to conftest
