+++ tests/test_metadata.py
@@ -1,7 +1,8 @@
+"""Tests for Metada class"""
+
 import logging
 import os
-import shutil
-from typing import Optional, Generator
+from typing import Optional
 from pathlib import Path
 from unittest.mock import patch
 
@@ -22,28 +23,6 @@ def metadata_model(helpers, data_model_labels):
     )
 
     return metadata_model
-
-
-@pytest.fixture
-def test_bulkrnaseq(helpers: Helpers) -> Generator[Path, None, None]:
-    """create temporary copy of test_BulkRNAseq.csv
-    This fixture creates a temporary copy of the original 'test_BulkRNAseq.csv' file
-    After test, the copied file is removed.
-    Args:
-        helpers (Helpers): Helpers fixture
-
-    Yields:
-        Generator[Path, None, None]: temporary file path of the copied version test_BulkRNAseq.csv
-    """
-    # original bulkrnaseq csv
-    original_test_path = helpers.get_data_path("mock_manifests/test_BulkRNAseq.csv")
-    # Copy the original CSV file to a temporary directory
-    temp_csv_path = helpers.get_data_path("mock_manifests/test_BulkRNAseq2.csv")
-    shutil.copyfile(original_test_path, temp_csv_path)
-    yield temp_csv_path
-    # Teardown
-    if os.path.exists(temp_csv_path):
-        os.remove(temp_csv_path)
 
 
 class TestMetadataModel:
